Fix handling of mouse & touchscreen on big endian.
There was a very insidious bug lurking in the EventHub implementation.
The kernel provides a capabilities bitmask as an array of _long_ and
not _char_. Now if you are lucky enough to have a LE target, all is
well, but no such luck on BE.

The proper solution is to have everything use longs, so just do that.