Fix non detected mouse.
Another scary bug. Things worked by chance in EventHub.
Another explanation needed. The way that bit arrays
were declared for passing to the ioctls were of the type
	uint8_t map[(MAP_MAX + 1) / 8];
with MAP_MAX typically
	#define MAP_MAX 0x0f
Can you spot the error? If the bit you tried to test
against fell in the _last_ 7 bits you would read garbage
(typically zero).

So we changed the uint8_t to unsigned long...
	unsigned long map[(MAP_MAX + 1) / 32];

size of map is now 0.

The correct solution is to use the form
(MAP_MAX + 31) / 32, and we use a nice define as well, since
this is quite ugly.